requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
datasets==3.6.0
dill==0.3.8
editdistpy==0.1.5
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
fsspec==2025.3.0
greenlet==3.2.1
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
inexactsearch==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.4.0
silpa_common==0.3
six==1.17.0
soundex==1.1.3
SQLAlchemy==2.0.40
sympy==1.14.0
symspellpy==6.7.7
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
